+++ Radamsa/vmf/src/modules/common/mutator/RadamsaReplaceLineMutator.cpp
@@ -156,8 +156,8 @@ void RadamsaReplaceLineMutator::mutateTestCase(StorageModule& storage, StorageEn
                             numLines);
     }
 
-    const size_t original_lineIndex = this->rand->randBetween(characterIndex, numLines - 1);
-    const size_t new_lineIndex = this->rand->randBetween(characterIndex, numLines - 2); // extra -1 because vector size is one less after orignal is removed
+    const size_t original_lineIndex = static_cast<size_t>(this->rand->randBetween(static_cast<unsigned long>(characterIndex), static_cast<unsigned long>(numLines - 1)));
+    const size_t new_lineIndex = static_cast<size_t>(this->rand->randBetween(static_cast<unsigned long>(characterIndex), static_cast<unsigned long>(numLines - 2))); // extra -1 because vector size is one less after orignal is removed
 
     // swap original with new
     const size_t original_value = lineOrder[original_lineIndex];
